+++ CHANGELOG.md
@@ -13,6 +13,13 @@ For older change log history see the [historic changelog](HISTORIC_CHANGELOG.md)
   - Added read-only value Enabled indicating whether a pre-staged account is Enabled or Disabled (Unoccupied).
 - `ActiveDirectoryDsc`
   - Added strings.psd1 for HQRM compliance.
+- `ADObjectPermissionEntry`
+  - The "ObjectType" parameter now supports the display name of the object to which
+    the access rule applies, in addition to the schema GUID.
+    ([issue #744](https://github.com/dsccommunity/ActiveDirectoryDsc/issues/744)).
+  - The "InheritedObjectType" parameter now supports the display name of the object
+    type that can inherit this access rule, in addition to the schema GUID.
+    ([issue #744](https://github.com/dsccommunity/ActiveDirectoryDsc/issues/744)).
 - ADDomain
   - Skip LCM reboot signal if `SuppressReboot` parameter is set to `true`
     ([issue #742](https://github.com/dsccommunity/ActiveDirectoryDsc/issues/742)).